Add MIT License
I like the MIT License because it's simple and ubiquitous. However, it
does mean that anyone's free to use/modify our software.

Justification:
--------------
I think this is a good thing given that our goal is greater awareness of
a major social/political problem. If someone wants to clone our repo and
make a new frontend with a different focus, that spreads the message
further. The MIT license also makes it clear that derivatives don't
necessarily have our blessing.

Alternatives:
-------------
However, if we're concerned with controlling what our (or Thursday
Network) is associated with or for some other reason don't want to share
code, we could use something else.

More license options:
  https://choosealicense.com/

About licensing a repo:
  https://help.github.com/articles/licensing-a-repository/